debug take off works nearly perfectly; stepping stone towards other variations of nv goggle take off | READ DESC FOR IMPORTANT ANIM INFO
So I learned something new just now. If you try to play a Roblox animation reversed (meaning you pass in a speed of -1) while the animation is still playing, it will appear as if the animation does not play for some reason. This was the problem I was facing when using animationTrack.Stopped. animation.Track ended was what I needed. Check documentation for these if you need clarification.