Establish Player Flow
Player flow being how the player experiences the game, going from the loading screen, to the title screen, to the main game, to the death screen.

Switched up from developing the CorpseLootable to this because a proper respawn system needs to be put into place before I can continue and finish the CorpseLootable system.

Since I had to do that, I thought in my head that I might as well establish the whole player flow to have something more robust and easier to refactor later.